backend/src/pyproject.toml
[tool.poetry]
name = "cometa"
version = "2.5.54"
description = ""
authors = ["AMVARA CONSULTING <info@amvara.de>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2.3"
psycopg2 = "^2.8.6"
psycopg2-binary = "^2.8.6"
djangorestframework = "^3.12.4"
tqdm = "^4.60.0"
django-filter = "^2.4.0"
coreapi = "^2.3.3"
requests = "^2.25.1"
django-cors-headers = "^3.7.0"
awesome-slugify = "^1.6.5"
bcrypt = "^3.2.0"
xhtml2pdf = "^0.2.5"
pycryptodome = "^3.10.1"
sentry-sdk = "^1.1.0"
Pillow = "^8.2.0"
stripe = "^2.56.0"
Pygments = "^2.9.0"
python-crontab = "^2.5.1"
gunicorn = "^20.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
